[REFACTOR] 타임라인 조회 로직 분리 #86
